requirements.txt
streamlit>=1.27.0
plotly>=5.16.0
pandas>=2.0.0
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=1.0.0
watchdog>=3.0.0
streamlit-extras>=0.3.0
streamlit-autorefresh>=0.1.0
APScheduler>=3.10.0
httpx>=0.24.0
pydantic>=2.0.0
streamlit-authenticator>=0.2.2
